# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AFRIGAZ",
        "description": "Documentation API AFRIGAZ",
        "contact": {
            "email": "winnersthec001.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/bulkPurchaseStoreData": {
            "post": {
                "tags": [
                    "Bulk"
                ],
                "summary": "Créer",
                "operationId": "a8526928e09135236eac008e63d3599a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "supplier_id"
                                ],
                                "properties": {
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "invoice_number": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "unit_price_per_kg": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity_kg": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "purchase_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-03"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Données créées avec succès"
                    },
                    "422": {
                        "description": "Validation des données échouée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
        "/api/v1/bulkPurchaseUpdate/{id}": {
            "put": {
                "tags": [
                    "Bulk"
                ],
                "summary": "Modifier",
                "operationId": "f5c31b352642cbf27b56dbfe8b4e4204",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "invoice_number": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "unit_price_per_kg": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity_kg": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "purchase_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-03"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "mise à jour"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                }
            }
        },
        "/api/v1/bulkPurchaseDelete/{id}": {
            "put": {
                "tags": [
                    "Bulk"
                ],
                "summary": "Supprimer",
                "operationId": "a29dfc39421949a3895e8fe4f2d69aff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimée"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                }
            }
        },
        "/api/v1/lostQuantityStore/{id}": {
            "put": {
                "tags": [
                    "Bulk"
                ],
                "summary": "Modifier",
                "operationId": "552a396967460dbe26a68680da40346a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lost_Quantity_kg": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                }
            }
        },
...
    },
...
}